Refactor `ExitAwareWebView` to become Google Ads's own specific webview `GoogleAdsWebvView`
