+++ WooCommerce/src/main/res/layout/fragment_googleads_webview.xml
