+++ WooCommerce/src/main/res/navigation/nav_graph_main.xml
@@ -514,10 +514,10 @@
         android:id="@+id/action_global_WPComWebViewFragment"
         app:destination="@id/WPComWebViewFragment" />
     <fragment
-        android:id="@+id/ExitAwareWebViewFragment"
-        android:name="com.woocommerce.android.ui.common.exitawarewebview.ExitAwareWebViewFragment"
-        android:label="ExitAwareWebViewFragment"
-        tools:layout="@layout/fragment_exitaware_webview">
+        android:id="@+id/GoogleAdsWebViewFragment"
+        android:name="com.woocommerce.android.ui.google.webview.GoogleAdsWebViewFragment"
+        android:label="GoogleAdsWebViewFragment"
+        tools:layout="@layout/fragment_googleads_webview">
         <argument
             android:name="urlToLoad"
             app:argType="string" />
@@ -538,19 +538,19 @@
         <argument
             android:name="displayMode"
             android:defaultValue="REGULAR"
-            app:argType="com.woocommerce.android.ui.common.exitawarewebview.ExitAwareWebViewViewModel$DisplayMode" />
+            app:argType="com.woocommerce.android.ui.google.webview.GoogleAdsWebViewViewModel$DisplayMode" />
         <argument
             android:name="urlComparisonMode"
             android:defaultValue="PARTIAL"
-            app:argType="com.woocommerce.android.ui.common.exitawarewebview.ExitAwareWebViewViewModel$UrlComparisonMode" />
+            app:argType="com.woocommerce.android.ui.google.webview.GoogleAdsWebViewViewModel$UrlComparisonMode" />
         <argument
             android:name="clearCache"
             android:defaultValue="false"
             app:argType="boolean" />
     </fragment>
     <action
-        android:id="@+id/action_global_ExitAwareWebViewFragment"
-        app:destination="@id/ExitAwareWebViewFragment" />
+        android:id="@+id/action_global_googleAdsWebViewFragment"
+        app:destination="@id/GoogleAdsWebViewFragment" />
 
     <action
         android:id="@+id/action_global_userEligibilityErrorFragment"
